requirements.txt
aiohttp==3.8.3
aiosignal==1.2.0
async-timeout==4.0.2
attrs==22.1.0
charset-normalizer==2.1.1
contourpy==1.0.5
cycler==0.11.0
fonttools==4.37.4
frozenlist==1.3.1
idna==3.4
kiwisolver==1.4.4
matplotlib==3.6.1
multidict==6.0.2
numpy==1.23.4
packaging==21.3
Pillow==9.2.0
pyparsing==3.0.9
python-dateutil==2.8.2
PyYAML==6.0
scipy==1.9.2
Shapely==1.8.5.post1
six==1.16.0
vtk==9.2.2
wslink==1.8.4
yarl==1.8.1
